+++ moesland-app/src/containers/NewsItemDetailView.js
