+++ moesland-app/src/containers/NewsItemListView.js
